Move send hello to status plugin
